+++ src/main/java/group3/africa/cropnest/dto/ProductRequestDTO.java
@@ -1,22 +1,21 @@
-package group3.africa.cropnest.Dtos;
+package group3.africa.cropnest.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class ProductResponseDTO {
-    private long productId;
+public class ProductRequestDTO {
     private String productName;
     private String productDescription;
     private double productPrice;
     private String productImageUrl;
     private double discountedPrice;
     private Integer productQuantity;
+    //    private Category categoryName;
 
 
-    public ProductResponseDTO(long productId, String productName, String productDescription, double productPrice, String productImageUrl, double discountedPrice, Integer productQuantity) {
-        this.productId = productId;
+    public ProductRequestDTO(String productName, String productDescription, double productPrice, String productImageUrl, double discountedPrice, Integer productQuantity) {
         this.productName = productName;
         this.productDescription = productDescription;
         this.productPrice = productPrice;
