+++ src/main/java/group3/africa/cropnest/dto/ProductResponseDTO.java
@@ -1,21 +1,22 @@
-package group3.africa.cropnest.Dtos;
+package group3.africa.cropnest.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class ProductRequestDTO {
+public class ProductResponseDTO {
+    private long productId;
     private String productName;
     private String productDescription;
     private double productPrice;
     private String productImageUrl;
     private double discountedPrice;
     private Integer productQuantity;
-    //    private Category categoryName;
 
 
-    public ProductRequestDTO(String productName, String productDescription, double productPrice, String productImageUrl, double discountedPrice, Integer productQuantity) {
+    public ProductResponseDTO(long productId, String productName, String productDescription, double productPrice, String productImageUrl, double discountedPrice, Integer productQuantity) {
+        this.productId = productId;
         this.productName = productName;
         this.productDescription = productDescription;
         this.productPrice = productPrice;
