+++ src/main/java/group3/africa/cropnest/exceptions/CategoryNotFoundException.java
@@ -1,4 +1,4 @@
-package group3.africa.cropnest.Exceptions;
+package group3.africa.cropnest.exceptions;
 
 public class CategoryNotFoundException extends RuntimeException {
     public CategoryNotFoundException(String message) {
